Improve image conversion task.
This handles geotiffs and multipage images.  It has support for built-in
compressors and for compressing with jp2k; when using jp2k COM markers
are removed to save a small amount of space.  Compression is
parallelized when possible.

The converter doesn't use subprocess (it uses the python interfacesfor
vips, GDAL, and glymur); this keeps the memory footprint smaller.

Move converter to utilities.  This will allow it to be used without
girder.